+++ MoMoney.Core/Models/DetailedStock.cs
@@ -3,14 +3,17 @@ namespace MoMoney.Core.Models;
 
 public partial class DetailedStock : Stock
 {
-    public decimal MarketValue => Symbol.EndsWith(":TSE") ? MarketPrice * Quantity : MarketPrice * Quantity * 1.35m;
+    public string FullName => $"{Symbol}:{Market}";
+    public decimal MarketValue => MarketPrice * Quantity;
     public decimal BookValue => Quantity * Cost;
     public decimal Change => MarketValue - BookValue;
     public decimal ChangePercent => (MarketValue / BookValue) - 1;
 
     public DetailedStock(Stock stock)
     {
+        StockID = stock.StockID;
         Symbol = stock.Symbol;
+        Market = stock.Market;
         Quantity = stock.Quantity;
         Cost = stock.Cost;
         MarketPrice = stock.MarketPrice;