+++ MoMoney.Core/Models/Stock.cs
@@ -5,8 +5,10 @@ namespace MoMoney.Core.Models;
 
 public partial class Stock : ObservableObject
 {
-    [PrimaryKey]
+    [PrimaryKey, AutoIncrement]
+    public int StockID { get; set; }
     public string Symbol { get; set; }
+    public string Market { get; set; }
     public decimal Quantity { get; set; }
     public decimal Cost { get; set; }
 
@@ -15,9 +17,10 @@ public partial class Stock : ObservableObject
 
     public Stock() { }
 
-    public Stock(string symbol, decimal quantity, decimal cost, decimal marketPrice)
+    public Stock(string symbol, string market, decimal quantity, decimal cost, decimal marketPrice)
     {
         Symbol = symbol;
+        Market = market;
         Quantity = quantity;
         Cost = cost;
         MarketPrice = marketPrice;
@@ -26,6 +29,7 @@ public partial class Stock : ObservableObject
     public Stock (Stock stock)
     {
         Symbol = stock.Symbol;
+        Market = stock.Market;
         Quantity = stock.Quantity;
         Cost = stock.Cost;
         MarketPrice = stock.MarketPrice;