+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Changed stock quantity from int to decimal to allow fractional shares
 * Updated look of Insights page
 * Condensed Add/Edit pages, made base viewmodel
+* Separated Stock symbol and market
 * Added component for CRUD buttons
 
 # MoMoney v1.2.0